Refactor Dot to create Person class
Refactor Dot class, split out person functions into new Person class.
